webusb-friendly labelling
